Fix issue where the currentScenario is not current
Angular UI-Router works by maintaining a graph of states, in this case the currentScenario only gets resolved when the page is initially loaded (due to our URLs). We now force a reload of the state whenever we save. This forces another GET request to the server, but it will maintain consistency. 

Please /do not/ touch the $scope.scenario stuff …